requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
Automat==20.2.0
autopep8==1.5.6
backcall==0.2.0
bleach==3.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
cryptography==3.3.1
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
dj-places==4.0.0
Django==3.1.3
django-allauth==0.44.0
django-crispy-forms==1.11.2
entrypoints==0.3
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.1
json2table==1.1.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.5.1
nltk==3.5
notebook==6.2.0
numpy==1.20.0
oauthlib==3.1.0
packaging==20.8
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.9.0
prompt-toolkit==3.0.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.7.4
PyHamcrest==2.0.2
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.0
python-ldap @ file:///C:/Users/mohil/Downloads/python_ldap-3.3.1-cp39-cp39-win_amd64.whl
python3-openid==3.2.0
pytz==2021.1
pywin32==300
pywinpty @ file:///C:/Users/mohil/Downloads/pywinpty-0.5.7-cp39-cp39-win_amd64.whl
pyzmq==21.0.1
qtconsole==5.0.2
QtPy==1.9.0
razorpay==1.2.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.5.0
simonspeckciphers==1.0.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
sqlparse==0.4.1
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.56.2
traitlets==4.3.3
tweepy==3.10.0
txaio==20.12.1
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zope.interface==5.2.0
